+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Lenstack Changelog
 
 ## [Current v1.8 - Lenstack]
-* New Dataset search api for better performance.
+* Added limit and offset to new datasetlibrary api
+* New Dataset datasetlibrary api for better performance.
 
 ## [v1.7 - Lenstack]
 * New Dataset query model replacing Fuzzy search for better performance.